chore: move create-s3-repo to examples folder
The `createRepo` convenience function has a dependency on `ipfs-repo`
which can create weird dependency loops.

Instead have it as an example of how to use this module then the user
can copy it to their codebase and also becomes in charge of which
version of `ipfs-repo` they use.
